moodcraft/apps/nlp-service/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1

# LangChain Ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# Vector Databases
pinecone-client>=5.0.0
qdrant-client>=1.12.0
chromadb>=0.5.0

# LLM Providers
openai>=1.50.0
anthropic>=0.39.0

# ML/NLP
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0
numpy>=2.0.0
scikit-learn>=1.5.0

# Database
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
prisma>=0.15.0

# Utilities
httpx>=0.27.0
aiohttp>=3.10.0
pyyaml>=6.0.0
tenacity>=9.0.0

# Observability
structlog>=24.0.0
prometheus-client>=0.21.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
httpx>=0.27.0
respx>=0.21.0
factory-boy>=3.3.0
